-correction de bugs -reste a faire le repo/controller/service pour groupe, concert, scene -configurer le pull de connexion pour lancer l'app

+++ src/main/java/com/hellmetz/festival/model/Style.java
@@ -21,12 +21,10 @@ public class Style {
     public Style() {}
 
     // Getters / Setters
-    public int getId() { return id; }
-
+    public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
 
     public String getLibelle() { return libelle; }
-
     public void setLibelle(String libelle) { this.libelle = libelle; }
 
     public String getDescription() { return description; }
